5379: move stackable material
