+++ src/apps/loan-list/list/loan-list-items.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect, useState } from "react";
 import { removeLoansWithDuplicateDueDate } from "../utils/helpers";
-import StackableMaterial from "../materials/stackable-material";
+import StackableMaterial from "../materials/stackable-material/stackable-material";
 import { GetMaterialManifestationQuery } from "../../../core/dbc-gateway/generated/graphql";
 import { ListView } from "../../../core/utils/types/list-view";
 import { LoanMetaDataType } from "../../../core/utils/helpers/LoanMetaDataType";
